migrate telemetry tasks to separate sdk
